requirements.txt
aiofiles==25.1.0
aiohappyeyeballs==2.6.1
aiohttp==3.13.2
aiohttp-retry==2.9.1
aiosignal==1.4.0
aiosqlite==0.21.0
alphashape==1.3.1
annotated-doc==0.0.4
annotated-types==0.7.0
anyio==4.12.0
argon2-cffi==25.1.0
argon2-cffi-bindings==25.1.0
attrs==25.4.0
beautifulsoup4==4.14.3
brotli==1.2.0
cachetools==6.2.2
certifi==2025.11.12
cffi==2.0.0
chardet==5.2.0
charset-normalizer==3.4.4
click==8.3.1
click-log==0.4.0
colorama==0.4.6
Crawl4AI==0.7.7
cryptography==46.0.3
cssselect==1.3.0
dataclasses-json==0.6.7
distro==1.9.0
dnspython==2.8.0
duckduckgo_search==8.1.1
ecdsa==0.19.1
email-validator==2.3.0
fake-http-header==0.3.5
fake-useragent==2.2.0
fastapi==0.123.0
fastapi-cli==0.0.16
fastapi-cloud-cli==0.5.2
fastar==0.8.0
fastuuid==0.14.0
filelock==3.20.0
filetype==1.2.0
frozenlist==1.8.0
fsspec==2025.10.0
google-ai-generativelanguage==0.9.0
google-api-core==2.28.1
google-auth==2.43.0
googleapis-common-protos==1.72.0
greenlet==3.2.4
groq==0.36.0
grpcio==1.76.0
grpcio-status==1.76.0
h11==0.16.0
h2==4.3.0
hpack==4.1.0
httpcore==1.0.9
httptools==0.7.1
httpx==0.28.1
httpx-sse==0.4.3
huggingface-hub==0.36.0
humanize==4.14.0
hyperframe==6.1.0
idna==3.11
importlib_metadata==8.7.0
Jinja2==3.1.6
jiter==0.12.0
joblib==1.5.2
jsonpatch==1.33
jsonpointer==3.0.0
jsonschema==4.25.1
jsonschema-specifications==2025.9.1
langchain==1.1.0
langchain-classic==1.0.0
langchain-community==0.4.1
langchain-core==1.1.0
langchain-google-genai==3.2.0
langchain-groq==1.1.0
langchain-openai==1.1.0
langchain-pinecone==0.2.13
langchain-text-splitters==1.0.0
langgraph==1.0.4
langgraph-checkpoint==3.0.1
langgraph-prebuilt==1.0.5
langgraph-sdk==0.2.10
langsmith==0.4.49
lark==1.3.1
litellm==1.80.5
lxml==5.4.0
markdown-it-py==4.0.0
MarkupSafe==3.0.3
marshmallow==3.26.1
mdurl==0.1.2
mpmath==1.3.0
multidict==6.7.0
mypy_extensions==1.1.0
networkx==3.6
nltk==3.9.2
numpy==2.3.5
openai==2.8.1
orjson==3.11.4
ormsgpack==1.12.0
packaging==24.2
patchright==1.56.0
pillow==12.0.0
pinecone==7.3.0
pinecone-client==6.0.0
pinecone-plugin-assistant==1.8.0
pinecone-plugin-interface==0.0.7
playwright==1.56.0
primp==0.15.0
propcache==0.4.1
proto-plus==1.26.1
protobuf==6.33.1
psutil==7.1.3
psycopg2-binary
pwdlib==0.3.0
pyasn1==0.6.1
pyasn1_modules==0.4.2
pycparser==2.23
pydantic==2.12.5
pydantic-settings==2.12.0
pydantic_core==2.41.5
pyee==13.0.0
Pygments==2.19.2
PyJWT==2.10.1
pyOpenSSL==25.3.0
python-dateutil==2.9.0.post0
python-dotenv==1.2.1
python-jose==3.5.0
python-multipart==0.0.20
PyYAML==6.0.3
rank-bm25==0.2.2
referencing==0.37.0
regex==2025.11.3
requests==2.32.5
requests-toolbelt==1.0.0
rich==14.2.0
rich-toolkit==0.17.0
rignore==0.7.6
rpds-py==0.30.0
rsa==4.9.1
rtree==1.4.1
safetensors==0.7.0
scikit-learn==1.7.2
scipy==1.16.3
sentence-transformers==5.1.2
sentry-sdk==2.46.0
setuptools==80.9.0
shapely==2.1.2
shellingham==1.5.4
simsimd==6.5.3
six==1.17.0
sniffio==1.3.1
snowballstemmer==2.2.0
soupsieve==2.8
SQLAlchemy==2.0.44
sqlmodel==0.0.27
starlette==0.50.0
sympy==1.14.0
tenacity==9.1.2
tf-playwright-stealth==1.2.0
threadpoolctl==3.6.0
tiktoken==0.12.0
tokenizers==0.22.1
torch==2.9.1
tqdm==4.67.1
transformers==4.57.3
trimesh==4.10.0
typer==0.20.0
typing-inspect==0.9.0
typing-inspection==0.4.2
typing_extensions==4.15.0
urllib3==2.5.0
uvicorn==0.38.0
watchfiles==1.1.1
websockets==15.0.1
xxhash==3.6.0
yarl==1.22.0
zipp==3.23.0
zstandard==0.25.0
